moving code, and making it beautifull
